Pull silent audio into a custom hook
